requirements.txt
# Core
python-dotenv>=1.0.0

# Backend (FastAPI)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
supabase>=2.3.0
psycopg2-binary>=2.9.0

# LLM & Agent (pinned for create_tool_calling_agent compatibility)
langchain>=0.1.0,<0.2.0
langchain-core>=0.1.52,<0.2.0
langchain-openai>=0.0.5,<0.1.0
langchain-community>=0.0.20

# Frontend
streamlit>=1.30.0
requests>=2.31.0
